Integrating KMeansBlocksImpl to KMeans
